SBERDOMA-962 add middleware
