+++ apps/condo/domains/common/utils/userid.utils.js
@@ -1,21 +1,12 @@
 import cookie from 'js-cookie'
-
-function makeId (length) {
-    let result = ''
-    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-    const charactersLength = characters.length
-    for (let i = 0; i < length; i++) {
-        result += characters.charAt(Math.floor(Math.random() * charactersLength))
-    }
-    return result
-}
+import { makeId } from './makeid.utils'
 
 function getCurrentUserId () {
-    let current = cookie.get('fingerprint')
+    let current = cookie.get('userId')
     if (!current) {
         current = makeId(12)
     }
-    cookie.set('fingerprint', current, { expires: 365 })
+    cookie.set('userId', current, { expires: 365 })
     return current
 }
