+++ apps/condo/domains/organization/schema/OrganizationEmployee.js
@@ -126,7 +126,7 @@ const OrganizationEmployee = new GQLListSchema('OrganizationEmployee', {
     },
     hooks: {
         validateInput: ({ resolvedData, existingItem, addValidationError, context }) => {
-            if (!hasRequestAndDbFields(['dv', { field: 'sender', checkCookies: true }], ['organization'], resolvedData, existingItem, context, addValidationError)) return
+            if (!hasRequestAndDbFields([{ field: 'dv', checkCookies: true }, { field: 'sender', checkCookies: true }], ['organization'], resolvedData, existingItem, context, addValidationError)) return
             if (!hasOneOfFields(['email', 'name', 'phone'], resolvedData, existingItem, addValidationError)) return
             const { dv } = resolvedData
             if (dv === 1) {
